+++ packages/contracts/test/deploy/managing-dao.ts
@@ -132,17 +132,6 @@ describe('Management DAO', function () {
       ),
     };
 
-    pluginsRepos = {
-      tokenVoting: PluginRepo__factory.connect(
-        hre.aragonPluginRepos.TokenVotingRepoProxy,
-        deployer
-      ),
-      addresslistVoting: PluginRepo__factory.connect(
-        hre.aragonPluginRepos.AddresslistVotingRepoProxy,
-        deployer
-      ),
-    };
-
     multisig = Multisig__factory.connect(
       hre.managementDAOMultisigPluginAddress,
       deployer
@@ -245,53 +234,6 @@ describe('Management DAO', function () {
           []
         )
       ).to.be.true;
-    });
-
-    context('Multisig Repo', async function () {
-      it('has permission to upgrade the TokenVoting PluginRepo', async function () {
-        expect(
-          await pluginsRepos.tokenVoting.isGranted(
-            pluginsRepos.tokenVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.UPGRADE_REPO_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-
-      it('has permission to maintain the repo', async function () {
-        expect(
-          await pluginsRepos.tokenVoting.isGranted(
-            pluginsRepos.tokenVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-    });
-
-    context('AddresslistVoting Repo', async function () {
-      it('has permission to upgrade the repo', async function () {
-        expect(
-          await pluginsRepos.addresslistVoting.isGranted(
-            pluginsRepos.addresslistVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.UPGRADE_REPO_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
-      it('has permission to maintain the repo', async function () {
-        expect(
-          await pluginsRepos.addresslistVoting.isGranted(
-            pluginsRepos.addresslistVoting.address,
-            managementDao.address,
-            PLUGIN_REPO_PERMISSIONS.MAINTAINER_PERMISSION_ID,
-            []
-          )
-        ).to.be.true;
-      });
     });
   });
 });