+++ packages/contracts/CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Deleted `AddresslistVoting`, `AddresslistVotingSetup` and related metadata.
+- Moved `TokenVoting`, `TokenVotingSetup` and related metadata to the [aragon/token-voting-plugin](https://github.com/aragon/token-voting-plugin) repo.
 - Moved `Multisig`, `MultisigSetup` and related metadata to the [aragon/multisig-plugin](https://github.com/aragon/multisig-plugin) repo.
 - Moved `Admin`, `AdminSetup` and related metadata to the [aragon/admin-plugin](https://github.com/aragon/admin-plugin) repo.
 - Moved